chore(mocker): give replay events a room
